cantina/checkers/checker2/requirements.txt
anyio==3.6.1
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==37.0.4
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
janus==1.0.0
msgpack==1.0.4
packaging==21.3
pycparser==2.21
pyparsing==3.0.9
python-can==4.0.0
PyYAML==6.0
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
typing_extensions==4.3.0
urllib3==1.26.12
wrapt==1.14.1
kaitaistruct
